intelli-claim-ai/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
langgraph
langsmith
pymongo
python-dotenv

pytesseract
pdf2image
pypdf
Pillow
opencv-python

spacy
rapidfuzz
pandas
numpy
scikit-learn

networkx
joblib